Se realizan correcciones correspondientes a EP1

+++ Carrito-D/Carrito-D/Models/Carrito.cs
@@ -6,6 +6,6 @@
         public bool Activo { get; set; }
         public Cliente Cliente { get; set; }
         public List<CarritoItem> CarritoItems { get; set; }
-        public double Subtotal { get; set; }
+        public float Subtotal { get; set; }
     }
 }
