+++ Carrito-D/Carrito-D/Models/Compra.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public Cliente Cliente { get; set; }
         public Carrito Carrito { get; set; }
-        public double Total { get; set; }
+        public float Total { get; set; }
         public DateTime Fecha { get; set; }
         public Sucursal Sucursal { get; set; }
